seta-auth/requirements.txt
flask==2.2.3
flask-cors==3.0.10
pymongo==4.3.3
pyjwt==2.6.0
gunicorn==20.1.0
flask_jwt_extended==4.4.4
python-secrets==22.6.1
flask-pymongo==2.3.0
flask-restx==1.0.5
python-interface==1.6.1
injector==0.20.1
flask-injector==0.14.0
requests==2.28.2
pytest==7.2.2
pycryptodome==3.17